requirements.txt
# MCP SDK
mcp>=0.3.0

# Snowflake dependencies
snowflake-connector-python>=3.0.0
snowflake-sqlalchemy>=1.4.0

# Database and data handling
pandas>=1.3.0
tabulate>=0.8.0
sqlalchemy>=1.4.0

# Utilities
python-dotenv>=0.19.0
tenacity>=8.0.0
typing-extensions>=4.0.0

# For development/debugging
pytest>=7.0.0
pytest-cov>=2.0.0
